Service fix-
Changes :

-[X] unused model removed
-[X] map function from json to model added
-[X] changes to use SelectedDetailProductData and SelectedProductData application models

+++ ChallengeiOSMeli/ChallengeiOSMeli/Clean/Domain/SelectedDetailProductData.swift
@@ -10,7 +10,7 @@ import Foundation
 struct SelectedDetailProductData{
     let imageUrl: String
     let productName: String
-    let productPrice: Int
-    let productOriginalPrice: Int
+    let productPrice: Double
+    let productOriginalPrice: Double
     let availableQuantity: Int
 }
